refact: move states related code from graphic to states
