+++ src/chart/boxplot/BoxplotView.ts
@@ -20,6 +20,7 @@
 import * as zrUtil from 'zrender/src/core/util';
 import ChartView from '../../view/Chart';
 import * as graphic from '../../util/graphic';
+import { setStatesStylesFromModel } from '../../util/states';
 import Path, { PathProps } from 'zrender/src/graphic/Path';
 import BoxplotSeriesModel, { BoxplotDataItemOption } from './BoxplotSeries';
 import GlobalModel from '../../model/Global';
@@ -180,7 +181,7 @@ function updateNormalBoxData(
 
     el.z2 = 100;
 
-    graphic.setStatesStylesFromModel(el, data.getItemModel<BoxplotDataItemOption>(dataIndex));
+    setStatesStylesFromModel(el, data.getItemModel<BoxplotDataItemOption>(dataIndex));
 }
 
 function transInit(points: number[][], dim: number, itemLayout: BoxplotItemLayout) {
