+++ src/chart/custom.ts
@@ -23,7 +23,8 @@ import {
     hasOwn, assert, isString, retrieve2, retrieve3, defaults, each, keys, isArrayLike, bind
 } from 'zrender/src/core/util';
 import * as graphicUtil from '../util/graphic';
-import * as SetLabelStyleOpt from '../label/labelStyle';
+import { enableElementHoverEmphasis, setAsHighDownDispatcher } from '../util/states';
+import * as labelStyleHelper from '../label/labelStyle';
 import {getDefaultLabel} from './helper/labelHelper';
 import createListFromArray from './helper/createListFromArray';
 import {getLayoutOnAxis} from '../layout/barGrid';
@@ -1113,11 +1114,11 @@ function updateElOnState(
             stateObj.textConfig = txCfgOpt;
         }
 
-        graphicUtil.enableElementHoverEmphasis(elDisplayable);
+        enableElementHoverEmphasis(elDisplayable);
     }
 
     if (isRoot) {
-        graphicUtil.setAsHighDownDispatcher(el, styleOpt !== false);
+        setAsHighDownDispatcher(el, styleOpt !== false);
     }
 }
 
@@ -1383,14 +1384,14 @@ function makeRenderItem(
         // Now that the feture of "auto adjust text fill/stroke" has been migrated to zrender
         // since ec5, we should set `isAttached` as `false` here and make compat in
         // `convertToEC4StyleForCustomSerise`.
-        const textStyle = SetLabelStyleOpt.createTextStyle(labelModel, null, opt, false, true);
+        const textStyle = labelStyleHelper.createTextStyle(labelModel, null, opt, false, true);
         textStyle.text = labelModel.getShallow('show')
             ? retrieve2(
                 customSeries.getFormattedLabel(dataIndexInside, NORMAL),
                 getDefaultLabel(data, dataIndexInside)
             )
             : null;
-        const textConfig = SetLabelStyleOpt.createTextConfig(textStyle, labelModel, opt, false);
+        const textConfig = labelStyleHelper.createTextConfig(textStyle, labelModel, opt, false);
 
         preFetchFromExtra(userProps, itemStyle);
         itemStyle = convertToEC4StyleForCustomSerise(itemStyle, textStyle, textConfig);
@@ -1415,7 +1416,7 @@ function makeRenderItem(
 
         let itemStyle = getItemStyleModel(dataIndexInside, EMPHASIS).getItemStyle();
         const labelModel = getLabelModel(dataIndexInside, EMPHASIS);
-        const textStyle = SetLabelStyleOpt.createTextStyle(labelModel, null, null, true, true);
+        const textStyle = labelStyleHelper.createTextStyle(labelModel, null, null, true, true);
         textStyle.text = labelModel.getShallow('show')
             ? retrieve3(
                 customSeries.getFormattedLabel(dataIndexInside, EMPHASIS),
@@ -1423,7 +1424,7 @@ function makeRenderItem(
                 getDefaultLabel(data, dataIndexInside)
             )
             : null;
-        const textConfig = SetLabelStyleOpt.createTextConfig(textStyle, labelModel, null, true);
+        const textConfig = labelStyleHelper.createTextConfig(textStyle, labelModel, null, true);
 
         preFetchFromExtra(userProps, itemStyle);
         itemStyle = convertToEC4StyleForCustomSerise(itemStyle, textStyle, textConfig);
@@ -1502,7 +1503,7 @@ function makeRenderItem(
     function font(
         opt: Parameters<typeof graphicUtil.getFont>[0]
     ): ReturnType<typeof graphicUtil.getFont> {
-        return SetLabelStyleOpt.getFont(opt, ecModel);
+        return labelStyleHelper.getFont(opt, ecModel);
     }
 }
