+++ src/chart/helper/Polyline.ts
@@ -18,6 +18,7 @@
 */
 
 import * as graphic from '../../util/graphic';
+import { enableHoverEmphasis } from '../../util/states';
 import type { LineDrawSeriesScope, LineDrawModelOption } from './LineDraw';
 import type List from '../../data/List';
 
@@ -72,7 +73,7 @@ class Polyline extends graphic.Group {
         const lineEmphasisState = line.ensureState('emphasis');
         lineEmphasisState.style = hoverLineStyle;
 
-        graphic.enableHoverEmphasis(this);
+        enableHoverEmphasis(this);
     };
 
     updateLayout(lineData: List, idx: number) {
