+++ src/component/legend/LegendView.ts
@@ -21,6 +21,7 @@ import {__DEV__} from '../../config';
 import * as zrUtil from 'zrender/src/core/util';
 import {createSymbol} from '../../util/symbol';
 import * as graphic from '../../util/graphic';
+import { enableHoverEmphasis } from '../../util/states';
 import {setLabelStyle, createTextStyle} from '../../label/labelStyle';
 import {makeBackground} from '../helper/listComponent';
 import * as layoutUtil from '../../util/layout';
@@ -319,7 +320,7 @@ class LegendView extends ComponentView {
                     defaultText: selectorItem.title
                 }
             );
-            graphic.enableHoverEmphasis(labelText);
+            enableHoverEmphasis(labelText);
         });
     }
 
@@ -455,7 +456,7 @@ class LegendView extends ComponentView {
 
         this.getContentGroup().add(itemGroup);
 
-        graphic.enableHoverEmphasis(itemGroup);
+        enableHoverEmphasis(itemGroup);
 
         // @ts-ignore
         itemGroup.__legendDataIndex = dataIndex;
